Move kraken specific tests to their own file
